+++ saleor/graphql/channel/types.py
@@ -1,15 +1,13 @@
-from typing import Union
+from typing import Type, Union, cast
 
 import graphene
 from django.db.models import Model
 from graphene.types.resolver import get_default_resolver
-from graphene_django import DjangoObjectType
 
 from ...channel import models
 from ...core.permissions import ChannelPermissions
-from ..core.connection import CountableDjangoObjectType
 from ..core.descriptions import ADDED_IN_31
-from ..core.types import CountryDisplay
+from ..core.types import CountryDisplay, ModelObjectType
 from ..decorators import permission_required
 from ..meta.types import ObjectWithMetadata
 from ..translations.resolvers import resolve_translation
@@ -40,7 +38,7 @@ class ChannelContextTypeForObjectType(graphene.ObjectType):
         return resolve_translation(root.node, info, language_code)
 
 
-class ChannelContextType(ChannelContextTypeForObjectType, DjangoObjectType):
+class ChannelContextType(ChannelContextTypeForObjectType, ModelObjectType):
     """A Graphene type that supports resolvers' root as ChannelContext objects."""
 
     class Meta:
@@ -50,10 +48,17 @@ class ChannelContextType(ChannelContextTypeForObjectType, DjangoObjectType):
     def is_type_of(cls, root: Union[ChannelContext, Model], info):
         # Unwrap node from ChannelContext if it didn't happen already
         if isinstance(root, ChannelContext):
-            return super().is_type_of(root.node, info)
+            root = cast(Model, root.node)
 
-        # Check type that was already unwrapped by the Entity union check
-        return super().is_type_of(root, info)
+        if isinstance(root, cls):
+            return True
+
+        if cls._meta.model._meta.proxy:
+            model = root._meta.model
+        else:
+            model = cast(Type[Model], root._meta.model._meta.concrete_model)
+
+        return model == cls._meta.model
 
 
 class ChannelContextTypeWithMetadataForObjectType(ChannelContextTypeForObjectType):
@@ -90,7 +95,13 @@ class ChannelContextTypeWithMetadata(
         abstract = True
 
 
-class Channel(CountableDjangoObjectType):
+class Channel(ModelObjectType):
+    id = graphene.GlobalID(required=True)
+    name = graphene.String(required=True)
+    is_active = graphene.Boolean(required=True)
+    slug = graphene.String(required=True)
+    currency_code = graphene.String(required=True)
+    slug = graphene.String(required=True)
     has_orders = graphene.Boolean(
         required=True, description="Whether a channel has associated orders."
     )
@@ -108,7 +119,6 @@ class Channel(CountableDjangoObjectType):
         description = "Represents channel."
         model = models.Channel
         interfaces = [graphene.relay.Node]
-        only_fields = ["id", "name", "slug", "currency_code", "is_active"]
 
     @staticmethod
     @permission_required(ChannelPermissions.MANAGE_CHANNELS)