+++ saleor/graphql/invoice/mutations.py
@@ -9,8 +9,8 @@ from ...invoice.notifications import send_invoice
 from ...order import events as order_events
 from ..core.mutations import ModelDeleteMutation, ModelMutation
 from ..core.types.common import InvoiceError
-from ..invoice.types import Invoice
 from ..order.types import Order
+from .types import Invoice
 from .utils import is_event_active_for_any_plugin
 
 
@@ -20,6 +20,7 @@ class InvoiceRequest(ModelMutation):
     class Meta:
         description = "Request an invoice for the order using plugin."
         model = models.Invoice
+        object_type = Invoice
         permissions = (OrderPermissions.MANAGE_ORDERS,)
         error_type_class = InvoiceError
         error_type_field = "invoice_errors"
@@ -121,6 +122,7 @@ class InvoiceCreate(ModelMutation):
     class Meta:
         description = "Creates a ready to send invoice."
         model = models.Invoice
+        object_type = Invoice
         permissions = (OrderPermissions.MANAGE_ORDERS,)
         error_type_class = InvoiceError
         error_type_field = "invoice_errors"
@@ -196,6 +198,7 @@ class InvoiceRequestDelete(ModelMutation):
     class Meta:
         description = "Requests deletion of an invoice."
         model = models.Invoice
+        object_type = Invoice
         permissions = (OrderPermissions.MANAGE_ORDERS,)
         error_type_class = InvoiceError
         error_type_field = "invoice_errors"
@@ -219,6 +222,7 @@ class InvoiceDelete(ModelDeleteMutation):
     class Meta:
         description = "Deletes an invoice."
         model = models.Invoice
+        object_type = Invoice
         permissions = (OrderPermissions.MANAGE_ORDERS,)
         error_type_class = InvoiceError
         error_type_field = "invoice_errors"
@@ -248,6 +252,7 @@ class InvoiceUpdate(ModelMutation):
     class Meta:
         description = "Updates an invoice."
         model = models.Invoice
+        object_type = Invoice
         permissions = (OrderPermissions.MANAGE_ORDERS,)
         error_type_class = InvoiceError
         error_type_field = "invoice_errors"
@@ -301,6 +306,7 @@ class InvoiceSendNotification(ModelMutation):
     class Meta:
         description = "Send an invoice notification to the customer."
         model = models.Invoice
+        object_type = Invoice
         permissions = (OrderPermissions.MANAGE_ORDERS,)
         error_type_class = InvoiceError
         error_type_field = "invoice_errors"