+++ saleor/graphql/translations/tests/test_translations.py
@@ -818,6 +818,10 @@ PRODUCT_TRANSLATE_MUTATION = """
                     }
                 }
             }
+            errors {
+                field
+                code
+            }
         }
     }
 """
@@ -922,6 +926,27 @@ def test_product_create_translation_by_translatable_content_id(
     assert data["product"]["translation"]["language"]["code"] == "PL"
 
 
+def test_product_create_translation_validates_name_length(
+    staff_api_client, product, permission_manage_translations
+):
+    product_id = graphene.Node.to_global_id("Product", product.id)
+    variables = {
+        "productId": product_id,
+        "input": {"description": None, "name": "Long" * 100},
+    }
+    response = staff_api_client.post_graphql(
+        PRODUCT_TRANSLATE_MUTATION,
+        variables,
+        permissions=[permission_manage_translations],
+    )
+    data = get_graphql_content(response)["data"]["productTranslate"]
+
+    assert data["product"] is None
+    assert data["errors"] == [
+        {"field": "name", "code": "INVALID"},
+    ]
+
+
 @freeze_time("1914-06-28 10:50")
 @patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_product_update_translation(
@@ -967,6 +992,10 @@ mutation productVariantTranslate(
                     code
                 }
             }
+        }
+        errors {
+            field
+            code
         }
     }
 }
@@ -1053,6 +1082,33 @@ def test_product_variant_update_translation(
     )
 
 
+def test_product_variant_translation_mutation_validates_inputs_length(
+    staff_api_client,
+    variant,
+    channel_USD,
+    permission_manage_translations,
+):
+    translatable_content_id = graphene.Node.to_global_id(
+        "ProductVariantTranslatableContent", variant.id
+    )
+    response = staff_api_client.post_graphql(
+        PRODUCT_VARIANT_TRANSLATE_MUTATION,
+        {
+            "productVariantId": translatable_content_id,
+            "input": {"name": "Wariant PL" * 100},
+        },
+        permissions=[permission_manage_translations],
+    )
+    data = get_graphql_content(response)["data"]["productVariantTranslate"]
+    assert data["productVariant"] is None
+    assert data["errors"] == [
+        {
+            "field": "name",
+            "code": "INVALID",
+        }
+    ]
+
+
 COLLECTION_TRANSLATE_MUTATION = """
 mutation collectionTranslate($collectionId: ID!, $input: TranslationInput!) {
     collectionTranslate(
@@ -1066,6 +1122,10 @@ mutation collectionTranslate($collectionId: ID!, $input: TranslationInput!) {
                     code
                 }
             }
+        }
+        errors {
+            field
+            code
         }
     }
 }
@@ -1187,6 +1247,25 @@ def test_collection_update_translation(
     mocked_webhook_trigger.assert_called_once_with(
         expected_payload, WebhookEventAsyncType.TRANSLATION_UPDATED
     )
+
+
+def test_collection_translation_mutation_validates_inputs_length(
+    staff_api_client, published_collection, permission_manage_translations
+):
+    collection_id = graphene.Node.to_global_id("Collection", published_collection.id)
+    description = dummy_editorjs("description", True)
+    response = staff_api_client.post_graphql(
+        COLLECTION_TRANSLATE_MUTATION,
+        {
+            "collectionId": collection_id,
+            "input": {"description": description, "name": "long" * 100},
+        },
+        permissions=[permission_manage_translations],
+    )
+    data = get_graphql_content(response)["data"]["collectionTranslate"]
+
+    assert data["collection"] is None
+    assert data["errors"] == [{"field": "name", "code": "INVALID"}]
 
 
 CATEGORY_TRANSLATE_MUTATION = """
@@ -1862,6 +1941,10 @@ SHIPPING_PRICE_TRANSLATE = """
                     }
                 }
             }
+            errors {
+                message
+                code
+            }
         }
     }
 """
@@ -1878,7 +1961,7 @@ def test_shipping_method_create_translation(
 ):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     shipping_method_id = graphene.Node.to_global_id(
-        "ShippingMethod", shipping_method.id
+        "ShippingMethodType", shipping_method.id
     )
     description = dummy_editorjs("description", True)
     variables = {
@@ -1959,7 +2042,7 @@ def test_shipping_method_update_translation(
     """
 
     shipping_method_id = graphene.Node.to_global_id(
-        "ShippingMethod", shipping_method.id
+        "ShippingMethodType", shipping_method.id
     )
     response = staff_api_client.post_graphql(
         query,
@@ -2088,6 +2171,27 @@ def test_shop_create_translation(
     )
 
 
+SHOP_SETTINGS_TRANSLATE_MUTATION = """
+    mutation shopSettingsTranslate($input: ShopSettingsTranslationInput!) {
+        shopSettingsTranslate(
+                languageCode: PL, input: $input) {
+            shop {
+                translation(languageCode: PL) {
+                    headerText
+                    language {
+                        code
+                    }
+                }
+            }
+            errors {
+                field
+                code
+            }
+        }
+    }
+"""
+
+
 @freeze_time("1914-06-28 10:50")
 @patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_shop_update_translation(
@@ -2103,24 +2207,10 @@ def test_shop_update_translation(
         language_code="pl", header_text="Nagłówek"
     )
 
-    query = """
-    mutation shopSettingsTranslate {
-        shopSettingsTranslate(
-                languageCode: PL, input: {headerText: "Nagłówek PL"}) {
-            shop {
-                translation(languageCode: PL) {
-                    headerText
-                    language {
-                        code
-                    }
-                }
-            }
-        }
-    }
-    """
-
     response = staff_api_client.post_graphql(
-        query, permissions=[permission_manage_translations]
+        SHOP_SETTINGS_TRANSLATE_MUTATION,
+        {"input": {"headerText": "Nagłówek PL"}},
+        permissions=[permission_manage_translations],
     )
     data = get_graphql_content(response)["data"]["shopSettingsTranslate"]
 
@@ -2133,6 +2223,28 @@ def test_shop_update_translation(
         expected_payload,
         WebhookEventAsyncType.TRANSLATION_UPDATED,
     )
+
+
+@freeze_time("1914-06-28 10:50")
+@patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
+def test_shop_translation_validates_values_lengths(
+    mocked_webhook_trigger,
+    staff_api_client,
+    site_settings,
+    permission_manage_translations,
+    settings,
+):
+    settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
+
+    response = staff_api_client.post_graphql(
+        SHOP_SETTINGS_TRANSLATE_MUTATION,
+        {"input": {"headerText": "Nagłówek PL" * 100}},
+        permissions=[permission_manage_translations],
+    )
+    data = get_graphql_content(response)["data"]["shopSettingsTranslate"]
+
+    assert data["shop"] is None
+    assert data["errors"] == [{"field": "headerText", "code": "INVALID"}]
 
 
 @pytest.mark.parametrize(
@@ -2558,7 +2670,7 @@ def test_translation_query_shipping_method(
     return_shipping_method,
 ):
     shipping_method_id = graphene.Node.to_global_id(
-        "ShippingMethod", shipping_method.id
+        "ShippingMethodType", shipping_method.id
     )
     perms = list(Permission.objects.filter(codename__in=perm_codenames))