+++ saleor/plugins/admin_email/tests/conftest.py
@@ -2,15 +2,16 @@ from unittest.mock import patch
 
 import pytest
 
+from ....plugins.models import EmailTemplate, PluginConfiguration
 from ...email_common import DEFAULT_EMAIL_VALUE
 from ...manager import get_plugins_manager
 from ..constants import (
     CSV_EXPORT_FAILED_DEFAULT_SUBJECT,
     CSV_EXPORT_FAILED_SUBJECT_FIELD,
     CSV_EXPORT_FAILED_TEMPLATE_FIELD,
-    CSV_PRODUCT_EXPORT_SUCCESS_DEFAULT_SUBJECT,
-    CSV_PRODUCT_EXPORT_SUCCESS_SUBJECT_FIELD,
-    CSV_PRODUCT_EXPORT_SUCCESS_TEMPLATE_FIELD,
+    CSV_EXPORT_SUCCESS_DEFAULT_SUBJECT,
+    CSV_EXPORT_SUCCESS_SUBJECT_FIELD,
+    CSV_EXPORT_SUCCESS_TEMPLATE_FIELD,
     SET_STAFF_PASSWORD_DEFAULT_SUBJECT,
     SET_STAFF_PASSWORD_SUBJECT_FIELD,
     SET_STAFF_PASSWORD_TEMPLATE_FIELD,
@@ -54,7 +55,7 @@ def admin_email_plugin(settings):
         csv_product_export_failed=DEFAULT_EMAIL_VALUE,
         set_staff_password_title=STAFF_ORDER_CONFIRMATION_DEFAULT_SUBJECT,
         staff_order_confirmation_title=SET_STAFF_PASSWORD_DEFAULT_SUBJECT,
-        csv_product_export_title=CSV_PRODUCT_EXPORT_SUCCESS_DEFAULT_SUBJECT,
+        csv_product_export_title=CSV_EXPORT_SUCCESS_DEFAULT_SUBJECT,
         csv_product_export_failed_title=CSV_EXPORT_FAILED_DEFAULT_SUBJECT,
         staff_password_reset_template=DEFAULT_EMAIL_VALUE,
         staff_password_reset_subject=STAFF_PASSWORD_RESET_DEFAULT_SUBJECT,
@@ -91,7 +92,7 @@ def admin_email_plugin(settings):
                             "value": staff_order_confirmation,
                         },
                         {
-                            "name": CSV_PRODUCT_EXPORT_SUCCESS_TEMPLATE_FIELD,
+                            "name": CSV_EXPORT_SUCCESS_TEMPLATE_FIELD,
                             "value": csv_product_export,
                         },
                         {
@@ -111,7 +112,7 @@ def admin_email_plugin(settings):
                             "value": set_staff_password_title,
                         },
                         {
-                            "name": CSV_PRODUCT_EXPORT_SUCCESS_SUBJECT_FIELD,
+                            "name": CSV_EXPORT_SUCCESS_SUBJECT_FIELD,
                             "value": csv_product_export_title,
                         },
                         {
@@ -125,3 +126,14 @@ def admin_email_plugin(settings):
         return manager.global_plugins[0]
 
     return fun
+
+
+@pytest.fixture
+def admin_email_template(admin_email_plugin):
+    plugin = admin_email_plugin()
+    config = PluginConfiguration.objects.get(identifier=plugin.PLUGIN_ID)
+    return EmailTemplate.objects.create(
+        name=STAFF_PASSWORD_RESET_TEMPLATE_FIELD,
+        value="Custom staff reset password email template",
+        plugin_configuration=config,
+    )