+++ saleor/plugins/avatax/__init__.py
@@ -218,18 +218,17 @@ def append_line_to_data(
 
 def append_shipping_to_data(
     data: List[Dict],
-    shipping_method_channel_listings: Optional["ShippingMethodChannelListing"],
+    shipping_price_amount: Optional[Decimal],
     shipping_tax_code: str,
 ):
     charge_taxes_on_shipping = (
         Site.objects.get_current().settings.charge_taxes_on_shipping
     )
-    if charge_taxes_on_shipping and shipping_method_channel_listings:
-        shipping_price = shipping_method_channel_listings.price
+    if charge_taxes_on_shipping and shipping_price_amount:
         append_line_to_data(
             data,
             quantity=1,
-            amount=shipping_price.amount,
+            amount=shipping_price_amount,
             tax_code=shipping_tax_code,
             item_code="Shipping",
         )
@@ -298,14 +297,20 @@ def get_checkout_lines_data(
                 ref2=line_info.variant.sku,
             )
 
-    append_shipping_to_data(
-        data, checkout_info.shipping_method_channel_listings, config.shipping_tax_code
-    )
+    delivery_method = checkout_info.delivery_method_info.delivery_method
+    if delivery_method:
+        price = getattr(delivery_method, "price", None)
+        append_shipping_to_data(
+            data,
+            price.amount if price else None,
+            config.shipping_tax_code,
+        )
+
     return data
 
 
 def get_order_lines_data(
-    order: "Order", config: AvataxConfiguration
+    order: "Order", config: AvataxConfiguration, invoice_transaction_type: bool = False
 ) -> List[Dict[str, Union[str, int, bool, None]]]:
     data: List[Dict[str, Union[str, int, bool, None]]] = []
     lines = order.lines.prefetch_related(
@@ -356,7 +361,11 @@ def get_order_lines_data(
             amount=undiscounted_amount,
         )
 
-        if undiscounted_amount != price_with_discounts_amount:
+        # for invoice transaction we want to include only final price
+        if (
+            not invoice_transaction_type
+            and undiscounted_amount != price_with_discounts_amount
+        ):
             append_line_to_data(
                 **append_line_to_data_kwargs,
                 amount=price_with_discounts_amount,
@@ -378,9 +387,12 @@ def get_order_lines_data(
     shipping_method_channel_listing = ShippingMethodChannelListing.objects.filter(
         shipping_method=order.shipping_method_id, channel=order.channel_id
     ).first()
-    append_shipping_to_data(
-        data, shipping_method_channel_listing, config.shipping_tax_code
-    )
+    if shipping_method_channel_listing:
+        append_shipping_to_data(
+            data,
+            shipping_method_channel_listing.price.amount,
+            config.shipping_tax_code,
+        )
     return data
 
 
@@ -506,12 +518,13 @@ def get_checkout_tax_data(
 
 def get_order_request_data(order: "Order", config: AvataxConfiguration):
     address = order.shipping_address or order.billing_address
-    lines = get_order_lines_data(order, config)
     transaction = (
         TransactionType.INVOICE
         if not (order.is_draft() or order.is_unconfirmed())
         else TransactionType.ORDER
     )
+    is_invoice_transaction = transaction == TransactionType.INVOICE
+    lines = get_order_lines_data(order, config, is_invoice_transaction)
     data = generate_request_data(
         transaction_type=transaction,
         lines=lines,