+++ saleor/plugins/tests/test_manager.py
@@ -554,23 +554,6 @@ def test_manager_apply_taxes_to_product(product, plugins, price, channel_USD):
     assert TaxedMoney(expected_price, expected_price) == taxed_price
 
 
-@pytest.mark.parametrize(
-    "plugins, price_amount",
-    [(["saleor.plugins.tests.sample_plugins.PluginSample"], "1.0"), ([], "10.0")],
-)
-def test_manager_apply_taxes_to_shipping(
-    shipping_method, address, plugins, price_amount, channel_USD
-):
-    shipping_price = shipping_method.channel_listings.get(
-        channel_id=channel_USD.id
-    ).price
-    expected_price = Money(price_amount, "USD")
-    taxed_price = PluginsManager(plugins=plugins).apply_taxes_to_shipping(
-        shipping_price, address, channel_slug=channel_USD.slug
-    )
-    assert TaxedMoney(expected_price, expected_price) == taxed_price
-
-
 def test_manager_sale_created(sale):
     plugins = ["saleor.plugins.tests.sample_plugins.PluginSample"]